Cache Spring AI vector store model name
