+++ apm-sniffer/apm-sdk-plugin/spring-plugins/spring-ai-1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/ai/v1/AbstractObservationVectorStoreInterceptor.java
@@ -29,13 +29,11 @@ import org.apache.skywalking.apm.agent.core.util.GsonUtil;
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 import org.apache.skywalking.apm.plugin.spring.ai.v1.contant.Constants;
 import org.springframework.ai.document.Document;
-import org.springframework.ai.embedding.EmbeddingOptions;
 import org.springframework.ai.vectorstore.SearchRequest;
 import org.springframework.ai.vectorstore.observation.AbstractObservationVectorStore;
 import org.springframework.ai.vectorstore.observation.VectorStoreObservationContext;
 import org.springframework.util.StringUtils;
 
-import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -104,52 +102,9 @@ public class AbstractObservationVectorStoreInterceptor implements InstanceMethod
     }
 
     private String resolveEmbeddingModelName(EnhancedInstance objInst) {
-        return resolveModelFromEmbeddingModel(objInst.getSkyWalkingDynamicField());
-    }
-
-    private String resolveModelFromEmbeddingModel(Object embeddingModel) {
-        if (embeddingModel == null) {
-            return null;
-        }
-        String model = resolveModelFromOptionsMethod(embeddingModel);
-        if (StringUtils.hasText(model)) {
-            return model;
-        }
-        model = resolveModelFromOptionsField(embeddingModel, "options");
-        if (StringUtils.hasText(model)) {
-            return model;
-        }
-        return resolveModelFromOptionsField(embeddingModel, "defaultOptions");
-    }
-
-    private String resolveModelFromOptionsMethod(Object embeddingModel) {
-        try {
-            Method method = embeddingModel.getClass().getMethod("getOptions");
-            return resolveModelFromOptions(method.invoke(embeddingModel));
-        } catch (Throwable ignored) {
-            return null;
-        }
-    }
-
-    private String resolveModelFromOptionsField(Object embeddingModel, String fieldName) {
-        Class<?> type = embeddingModel.getClass();
-        while (type != null) {
-            try {
-                Field field = type.getDeclaredField(fieldName);
-                field.setAccessible(true);
-                return resolveModelFromOptions(field.get(embeddingModel));
-            } catch (NoSuchFieldException e) {
-                type = type.getSuperclass();
-            } catch (Throwable ignored) {
-                return null;
-            }
-        }
-        return null;
-    }
-
-    private String resolveModelFromOptions(Object options) {
-        if (options instanceof EmbeddingOptions) {
-            return ((EmbeddingOptions) options).getModel();
+        Object context = objInst.getSkyWalkingDynamicField();
+        if (context instanceof VectorStoreEnhanceContext) {
+            return ((VectorStoreEnhanceContext) context).getEmbeddingModelName();
         }
         return null;
     }
